% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{activesession_backend}
\alias{activesession_backend}
\title{Retrieve an activesession backend.}
\usage{
activesession_backend(name, ...)
}
\arguments{
\item{name}{character. Name of the activesession backend. Currently,
supported values are \code{c("r", "file", "s3mpi", "cacher")}, but the
latter two require additional packages.}
}
\value{
An \code{activesession_backend} R6 object.
}
\description{
Retrieve an activesession backend.
}
