% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-R6.R
\docType{data}
\name{activesession_backend_R6}
\alias{activesession_backend_R6}
\title{An activesession backend R6 class template.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
activesession_backend_R6
}
\description{
An activesession backend R6 class template.
}
\keyword{datasets}
