% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activesession.R
\name{$.activesession}
\alias{$.activesession}
\title{Read from an activesession.}
\usage{
\method{$}{activesession}(object, el)
}
\arguments{
\item{object}{activesession. The activesession object.}

\item{el}{character. The accessed key.}
}
\description{
Read from an activesession.
}
