% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activesession.R
\name{$<-.activesession}
\alias{$<-.activesession}
\title{Write to an activesession.}
\usage{
\method{$}{activesession}(object, el) <- value
}
\arguments{
\item{object}{activesession. The activesession object.}

\item{el}{character. The written key.}

\item{value}{ANY. The value to assign.}
}
\description{
Write to an activesession.
}
