% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.directory}
\alias{is.directory}
\title{Determine whether the given path is a valid directory.}
\usage{
is.directory(path)
}
\arguments{
\item{path}{character. The path to check.}
}
\value{
\code{TRUE} or \code{FALSE} according as \code{path} is a valid
  existent directory.
}
\description{
Determine whether the given path is a valid directory.
}
