% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.simple_string}
\alias{is.simple_string}
\title{Checks if a string is a simple string.}
\usage{
is.simple_string(string)
}
\description{
A simple string is (a) character, (b) length 1, and (c) not blank.
}
